Data type support, lengthless column, raw get method
Support for CHAR, TINYINT, SMALLINT, BIGINT (long), BOOLEAN, FLOAT, and DOUBLE
Ability to omit length in columns by assigning a null value to length in the constructor. Useful for BOOLEAN and numeric data types

A new get method that returns the raw ResultSet instead of a processed object value. Useful if multiple data columns of the player are stored